packages/casbin-fastapi-decorator-jwt/pyproject.toml
[project]
name = "casbin-fastapi-decorator-jwt"
version = "0.1.0"
description = "JWT user provider for casbin-fastapi-decorator"
requires-python = ">=3.10"
dependencies = [
    "casbin-fastapi-decorator",
    "pyjwt>=2.0.0",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
casbin-fastapi-decorator = { workspace = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py", "act.py", "error.py"]
markers = [
    "integration: integration tests (real I/O, HTTP)",
    "unit: unit tests (no real I/O)",
    "jwt_provider: JWTUserProvider component tests",
]

[tool.bandit]
exclude_dirs = ["tests"]

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.0",
]